added font awesome. Created icons at the top of the calendar. Moved the AddClass modal to the calendar instead of the sidebar.
