+++ src/Components/Calendar/calendar.scss
@@ -8,7 +8,38 @@
   margin-bottom: 50px;
   top: 50px;
   left: 0;
+}
+.calendar {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
   border: solid 1px $gray;
+}
+.calendar-header {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+}
+.calendar-header-icon-wrap {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  margin-left: 20px;
+  margin-bottom: 10px;
+  cursor: pointer;
+  color: $green1;
+  transition: ease-in-out 0.2s;
+  &:hover {
+    color: $green2;
+  }
+  p {
+    font-size: 8pt;
+    font-weight: 600;
+    font-family: $Raleway;
+    margin: 5px 0px 0px 0px;
+  }
 }
 .dayname-row {
   width: 100%;
