+++ src/Components/Sidebar/sidebar.scss
@@ -87,233 +87,6 @@
 .css-yk16xz-control {
   border-radius: 0px !important;
 }
-.add-new-class-btn {
-  width: 100%;
-  height: 40px;
-  background: white;
-  border: solid 2px $blue;
-  border-radius: 10px;
-  cursor: pointer;
-  transition: ease-in-out 0.3s;
-  margin-top: 15px;
-  font-size: 12pt;
-  font-family: $Raj;
-  font-weight: 600;
-  &:hover {
-    background: $blue;
-    color: white;
-  }
-}
-.add-class-modal {
-  width: 60vw;
-  border: none;
-  outline: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.ReactModal__Overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: #00000030 !important;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-}
-.add-class-modal-wrap {
-  width: 60vw;
-  background: white;
-  border-radius: 5px;
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 20px;
-}
-.add-class-modal-header {
-  width: 100%;
-  height: 60px;
-  background: $green1;
-  color: white;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  border-radius: 5px 5px 0px 0px;
-  h2 {
-    margin: 0;
-    margin-left: 25px;
-    font-weight: 400;
-    font-family: $Raj;
-  }
-  button {
-    background: none;
-    border: none;
-    padding: 0px 5px;
-    margin: 0;
-    font-size: 18px;
-    font-weight: 600;
-    font-family: $Raj;
-    cursor: pointer;
-    transition: ease 0.3s;
-    &:hover {
-      box-shadow: 0px 0px 5px rgba(0, 0, 0, 0.2);
-      border-radius: 5px;
-      color: $green1;
-    }
-  }
-}
-.add-class-section-information {
-  width: 95%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-.section-info {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-}
-.section-info-left {
-  width: 48%;
-  display: flex;
-  flex-direction: column;
-  label {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    margin-bottom: 8px;
-    font-size: 10pt;
-    input {
-      width: 65%;
-      height: 25px;
-      margin-left: 8px;
-      border-radius: 3px;
-      border: solid 1px $lightGray;
-      background: none;
-      font-size: 10pt;
-    }
-  }
-}
-.section-info-right {
-  width: 48%;
-  display: flex;
-  flex-direction: column;
-  label {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    margin-bottom: 8px;
-    font-size: 10pt;
-    input {
-      width: 65%;
-      height: 25px;
-      margin-left: 8px;
-      border-radius: 3px;
-      border: solid 1px $lightGray;
-      background: none;
-      font-size: 10pt;
-    }
-  }
-}
-.section-course-attributes {
-  width: 95%;
-  display: flex;
-  justify-content: space-between;
-  margin-top: 10px;
-}
-.section-course-attribute {
-  width: 48%;
-  h3 {
-    width: 100%;
-    margin: 0px 0px 10px 0px;
-  }
-  textarea {
-    width: 100%;
-  }
-}
-.schedule-section {
-  width: 95%;
-  display: flex;
-  justify-content: space-between;
-}
-.schedule-section-column {
-  width: 30%;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  h3 {
-    width: 100%;
-    margin: 20px 0px 10px 0px;
-  }
-  input {
-    width: 98%;
-    height: 25px;
-    border-radius: 3px;
-    border: solid 1px $lightGray;
-    background: none;
-    font-size: 10pt;
-  }
-}
-.enrollment-section-text {
-  width: 95%;
-  display: flex;
-  justify-content: space-between;
-}
-.enrollment-section-text-column {
-  width: 30%;
-  h3 {
-    width: 100%;
-    border-bottom: solid 1px $lightGray;
-    margin: 20px 0px 10px 0px;
-  }
-  textarea {
-    width: 100%;
-  }
-}
-.enrollment-section {
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  label {
-    display: flex;
-    flex-direction: column;
-    input {
-      width: 80px;
-    }
-  }
-}
-.add-class-btns {
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 30px;
-  button {
-    width: 150px;
-    height: 35px;
-    border: none;
-    font-family: $Raj;
-    font-size: 12pt;
-    // margin-left: 20px;
-    cursor: pointer;
-    transition: ease-in-out 0.2s;
-  }
-  .add-class-cancel-btn {
-    background: $lightGray;
-    &:hover {
-      background: darken($lightGray, 10%);
-    }
-  }
-  .add-class-save-btn {
-    background: $green1;
-    color: white;
-    &:hover {
-      background: lighten($green1, 10%);
-    }
-  }
-}
 
 .reset-calendar {
   width: 100%;
